Split Uroboro.Tree in three modules.
